Streamba/photo-map-app/backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2
python-dotenv
